+++ src/components/SearchBar.tsx
@@ -4,10 +4,12 @@ import styles from "./SearchBar.module.css";
 
 const SearchBar = ({
   setSearch,
+  value,
   initialValue,
   placeholder,
 }: {
   setSearch;
+  value: string;
   initialValue?: string;
   placeholder?: string;
 }) => {
@@ -15,7 +17,7 @@ const SearchBar = ({
   return (
     <input
       type="text"
-      value={initialValue}
+      value={value ?? initialValue}
       placeholder={placeholder ?? t("skills.searchPlaceholder")}
       onChange={(e) => setSearch(e?.target?.value || "")}
       className={`bg-light-light dark:bg-dark-light w-full rounded-full p-4 bg-right ${styles.search}`}
